Separating Scala 2.12 and Scala 2.11 validation tests
Testing Scala 2.12.12 and 2.11.12.
Added Scala tests for different Java 10 and Java 12 bytecode profiles.

More info:
https://docs.scala-lang.org/overviews/jdk-compatibility/overview.html
