+++ org.jacoco.core.test.validation.scala2_11/src/org/jacoco/core/test/validation/scala2_11/ScalaSynchronizedTest.java
@@ -1,0 +1,27 @@
+/*******************************************************************************
+ * Copyright (c) 2009, 2020 Mountainminds GmbH & Co. KG and Contributors
+ * This program and the accompanying materials are made available under
+ * the terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *    Evgeny Mandrikov - initial API and implementation
+ *
+ *******************************************************************************/
+package org.jacoco.core.test.validation.scala;
+
+import org.jacoco.core.test.validation.ValidationTestBase;
+import org.jacoco.core.test.validation.scala2_11.targets.ScalaSynchronizedTarget;
+
+/**
+ * Test of synchronized block.
+ */
+public class ScalaSynchronizedTest extends ValidationTestBase {
+
+	public ScalaSynchronizedTest() {
+		super(ScalaSynchronizedTarget.class);
+	}
+
+}
