+++ org.jacoco.core.test.validation.scala2_11/src/org/jacoco/core/test/validation/scala2_11/targets/ScalaSynchronizedTarget.scala
@@ -7,25 +7,27 @@
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
- *    Gergely Fábián - initial API and implementation
+ *    Evgeny Mandrikov - initial API and implementation
  *
  *******************************************************************************/
-package org.jacoco.core.test.validation.scala.targets
+package org.jacoco.core.test.validation.scala2_11.targets
 
-import org.jacoco.core.test.validation.targets.Stubs.{exec, noexec, nop}
+import org.jacoco.core.test.validation.targets.Stubs.nop
 
 /**
- * Test target for anonymous functions.
+ * Test target for synchronized block.
  */
-object ScalaCaseClassTarget {
+object ScalaSynchronizedTarget {
 
-  case class Foo(a: Int, b: String, c: Set[String], d: Double) // assertPartlyCovered(31, 0)
+  private val lock: Object = new Object()
 
   def main(args: Array[String]): Unit = {
-
-    val foo = Foo(1, "test", Set.empty[String], 2.3) // assertFullyCovered()
-    val message = "" + foo.a + foo.b + foo.c.mkString(",") + foo.d // assertFullyCovered()
-
+    nop()
+    lock.synchronized { // assertFullyCovered()
+      /* Without filter next line covered partly */
+      nop() // assertFullyCovered()
+    } // assertEmpty()
+    nop()
   }
 
 }
