+++ org.jacoco.core.test.validation.scala2_12/src/org/jacoco/core/test/validation/scala2_12/ScalaCaseClassTest.java
@@ -7,21 +7,21 @@
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
- *    Evgeny Mandrikov - initial API and implementation
+ *    Gergely Fábián - initial API and implementation
  *
  *******************************************************************************/
 package org.jacoco.core.test.validation.scala;
 
 import org.jacoco.core.test.validation.ValidationTestBase;
-import org.jacoco.core.test.validation.scala.targets.ScalaSingletonObjectTarget;
+import org.jacoco.core.test.validation.scala2_12.targets.ScalaCaseClassTarget;
 
 /**
- * Test of singleton object.
+ * Test of synchronized block.
  */
-public class ScalaSingletonObjectTest extends ValidationTestBase {
+public class ScalaCaseClassTest extends ValidationTestBase {
 
-	public ScalaSingletonObjectTest() {
-		super(ScalaSingletonObjectTarget.class);
+	public ScalaCaseClassTest() {
+		super(ScalaCaseClassTarget.class);
 	}
 
 }
