+++ org.jacoco.core.test.validation.scala2_12/src/org/jacoco/core/test/validation/scala2_12/targets/ScalaSingletonObjectTarget.scala
@@ -1,0 +1,23 @@
+/*******************************************************************************
+ * Copyright (c) 2009, 2020 Mountainminds GmbH & Co. KG and Contributors
+ * This program and the accompanying materials are made available under
+ * the terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *    Evgeny Mandrikov - initial API and implementation
+ *
+ *******************************************************************************/
+package org.jacoco.core.test.validation.scala2_12.targets
+
+/**
+ * This test target is singleton object.
+ */
+object ScalaSingletonObjectTarget { // assertFullyCovered()
+
+  def main(args: Array[String]): Unit = { // assertFullyCovered()
+  } // assertEmpty()
+
+}
